backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
openai==1.12.0
slack-bolt==1.18.0
python-dotenv==1.0.0
pytest==9.0.2
pytest-asyncio==1.3.0
tenacity==8.2.3
